fix: geo API COUNT(*)::int OVER → (COUNT(*) OVER (...))::int cast 위치 수정
PostgreSQL은 집계함수에 cast 후 OVER 키워드를 허용하지 않음 — cast를 window 결과에 적용

+++ scripts/lib/api_routes/listings.mjs
@@ -939,9 +939,9 @@ export async function handleListingsGeo(req, res) {
             PARTITION BY _lat4, _lng4, rent_amount, deposit_amount
             ORDER BY _plat_pri, listing_id
           ) AS _cross_rk,
-          COUNT(*)::int OVER (
+          (COUNT(*) OVER (
             PARTITION BY _lat4, _lng4, rent_amount, deposit_amount
-          ) AS group_count,
+          ))::int AS group_count,
           ARRAY_AGG(listing_id) OVER (
             PARTITION BY _lat4, _lng4, rent_amount, deposit_amount
             ORDER BY _plat_pri, listing_id
